usage-service/airflow/requirements.txt
confluent-kafka==2.6.0
fastavro==1.12.1

apache-airflow==2.10.2
apache-airflow-providers-standard==1.3.0

python-dotenv==1.2.2
SQLAlchemy==1.4.52
psycopg2-binary==2.9.11

chromadb==0.5.15

langchain==1.2.15
langchain-core==1.2.29
langchain-community==0.4.1
langchain-openai==1.1.10
openai==2.31.0 
